Subject: nroff/troff 4.4BSD mdoc manpage format Index:	 src/usr.bin/roff/* src/ucb/more.c src/share/{termcap,tmac} src/man/man1/nroff.1 src/man/man7/mdoc.*
Description:
Repeat-By:

	These sections are replaced with a "few words" from the author of
	the changes: Anders Magnusson <ragge@tethuvudet.se>

*****************************************************
Here comes next large patch - this time is it regarding text and troff :-)
It's focused in getting the 4.4BSD mdoc man page format to work but there
are other benefits as well.

Notes about Troff:
The changes below did cut between 20-40% of the execution time of nroff,
tested on my 11/83.  Also, data size needed is decreased from 44k to 30k,
but text increased 4k.

User-visible changes:
====================
Use text tab files instead of relying on a specific compiler and binary format.
Tabfiles now have suffix .tab (for example) tab37.tab for TTY37.

The tab file layout is the same as what ditroff uses.

Search path for term and tab files (in order):
        dirs given in -M (tmac) or -F (tab)
        environment variable TERMTAB_PATH/TMAC_PATH
        system directories (from _PATH_TERMTAB/_PATH_TMAC)
        current directory.

    this is consistent with groff (and a superset of ditroff).

Remove the command line use of -num and +num; they are redundant (-o) and
neither is in groff nor ditroff.

Remove the .li request. It does not exist in any docs and is not in either
ditroff or groff.

Internal changes:
================
Use a builtin dumb printer definition similar to what groff (grotty) uses,
so if -T is not given no file will be read.

Allow for use of more than 64k of the temp file.  This code existed but
was disabled.  Just enabling it overflowed the data segment even when
using split I/D. It was using long for offset pointers which was remarkably
slow.  Changed to use ints and buffer numbers instead.

Store env in a struct instead of relying on compiler behaviour.

Decrease default input and output buffer to 1k (gives no change in execution
time).  This also fixes a bug where .cf would segfault (cannot ever have
been tested on 2BSD).

Use a hash table to improve search time for macros.

Implement a 1-way associative read cache for the temp file.

Use the temp file to save data from input in case of .so and reading from stdin.

Use stdio routines for output/error instead of homegrown routines.
*****************************************************

Fix:
	These 10 new files are added to the system:

./usr/src/man/man7/mdoc.7
./usr/src/man/man7/mdoc.samples.7
./usr/src/share/tmac/doc
./usr/src/share/tmac/doc-common
./usr/src/share/tmac/doc-ditroff
./usr/src/share/tmac/doc-nroff
./usr/src/share/tmac/doc-syms
./usr/src/share/tmac/strip.sed
./usr/src/usr.bin/roff/TODO
./usr/src/usr.bin/roff/nroff_term/mktab.c

	These 32 files are updated/overwritten. If you have local changes
	that you want to retain please make a backup before installing the
	update:

./VERSION
./usr/include/paths.h
./usr/src/man/man1/nroff.1
./usr/src/man/man7/Makefile
./usr/src/share/termcap/termcap.src
./usr/src/share/tmac/Makefile
./usr/src/share/tmac/tmac.an
./usr/src/ucb/more/more.c
./usr/src/usr.bin/roff/common_source/d.h
./usr/src/usr.bin/roff/common_source/hytab.c
./usr/src/usr.bin/roff/common_source/n1.c
./usr/src/usr.bin/roff/common_source/n2.c
./usr/src/usr.bin/roff/common_source/n3.c
./usr/src/usr.bin/roff/common_source/n4.c
./usr/src/usr.bin/roff/common_source/n5.c
./usr/src/usr.bin/roff/common_source/n7.c
./usr/src/usr.bin/roff/common_source/n8.c
./usr/src/usr.bin/roff/common_source/n9.c
./usr/src/usr.bin/roff/common_source/ni.c
./usr/src/usr.bin/roff/common_source/nii.c
./usr/src/usr.bin/roff/common_source/sdef.h
./usr/src/usr.bin/roff/common_source/suftab.c
./usr/src/usr.bin/roff/common_source/tdef.h
./usr/src/usr.bin/roff/common_source/tw.h
./usr/src/usr.bin/roff/nroff/n10.c
./usr/src/usr.bin/roff/nroff/n6.c
./usr/src/usr.bin/roff/nroff/ntab.c
./usr/src/usr.bin/roff/nroff_term/Makefile
./usr/src/usr.bin/roff/nroff_term/README
./usr/src/usr.bin/roff/troff/t10.c
./usr/src/usr.bin/roff/troff/t6.c
./usr/src/usr.bin/roff/troff/tab3.c

	To install this update save this entire patch file to a temporary
	file (/tmp/foo) and then:

cd /tmp